pyproject.toml
[tool.poetry]
name = "to-do-list"
version = "0.1.0"
description = ""
authors = ["GabrielRehbein <gabrielrehbei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.15.2"
django = "^5.1.5"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
